Console: Update rrd4j to version 3.9

See merge request i2p-hackers/i2p.i2p!178

+++ apps/jrobin/java/src/org/rrd4j/core/Archive.java
@@ -143,7 +143,7 @@ public class Archive implements RrdUpdater<Archive> {
             }
         }
         // update robin in bulk
-        int bulkUpdateCount = (int) Math.min(numUpdates / steps.get(), (long) rows.get());
+        int bulkUpdateCount = (int) Math.min(numUpdates / steps.get(), rows.get());
         robin.bulkStore(value, bulkUpdateCount);
         // update remaining steps
         long remainingUpdates = numUpdates % steps.get();
